rename TypesTop to ActivityTopTypesByPeriod
This is more in line with other naming and prepares us for the next step
of supporting multiple periods.
