+++ web/src/components/ActivityTopTypesByPeriod.svelte
