+++ web/src/components/TabActivity.svelte
@@ -3,7 +3,7 @@
   import ActivityAircraftByPeriod from "./ActivityAircraftByPeriod.svelte";
   import MetricAircraftSeen from "./MetricAircraftSeen.svelte";
   import MetricFlightsSeen from "./MetricFlightsSeen.svelte";
-  import TypesTop from './TypesTop.svelte';
+  import ActivityTopTypesByPeriod from './ActivityTopTypesByPeriod.svelte';
 </script>
 
 <div class="grid grid-cols-1 lg:grid-cols-1 xl:grid-cols-2 mt-10 gap-6">
@@ -20,6 +20,6 @@
     <ActivityAircraftByPeriod />
   </div>
   <div class="col">
-    <TypesTop />
+    <ActivityTopTypesByPeriod />
   </div>
 </div>
